added role isAdmin on user when created a product

+++ src/products/products.service.ts
@@ -1,16 +1,24 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { randomUUID } from 'crypto';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { handleErrorConstraintUnique } from 'src/utils/handle-error-unique.util';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { Product } from './entities/product.entity';
+import { User } from 'src/users/entities/user.entity';
 
 @Injectable()
 export class ProductsService {
   constructor(private readonly prisma: PrismaService) {}
 
-  async create(dto: CreateProductDto): Promise<Product> {
+  async create(dto: CreateProductDto, user: User): Promise<Product> {
+    if (!user.role) {
+      throw new UnauthorizedException();
+    }
     const id = randomUUID();
     const data = { ...dto, id };
 
@@ -35,7 +43,15 @@ export class ProductsService {
     return product;
   }
 
-  async update(id: string, dto: UpdateProductDto): Promise<Product> {
+  async update(
+    id: string,
+    dto: UpdateProductDto,
+    user: User,
+  ): Promise<Product> {
+    if (!user.role) {
+      throw new UnauthorizedException();
+    }
+
     await this.findOne(id);
 
     return await this.prisma.products
@@ -43,7 +59,11 @@ export class ProductsService {
       .catch(handleErrorConstraintUnique);
   }
 
-  async remove(id: string) {
+  async remove(id: string, user: User) {
+    if (!user.role) {
+      throw new UnauthorizedException();
+    }
+
     await this.findOne(id);
 
     return await this.prisma.users.delete({ where: { id } });
